requirements.txt
python-dotenv
colorama
reportlab
flask
pandas
